Add methods for account service and dependencies for junit

+++ src/main/java/com/webapp/banca/demo/model/entity/ClientEntity.java
@@ -21,7 +21,7 @@ public class ClientEntity {
     private String pLastName;
     private String pPhoneNumber;
     private int documentType;
-    private int documentNumber;
+    private Long documentNumber;
     private int pAge;
     private String pEmeal;
     private String pStatusCivil;
